Change type of pointer to data of input ANG record from double to float
The record in the file has FST type 'E' and data_bits 32.

   NOMV TV   ETIQUETTE        NI      NJ   ...  DTY   G
6- ANG  P@ WEIGHTSRC2        362     292   ...  E 32  X

+++ src/utils/georef_cstintrp_reindexer.c
@@ -75,7 +75,7 @@ int ReIndex(char **In,char *Out,char* FromTo,int *OtherDims,int BDW, int Orca) {
             if (in) {
                data[v++]=i;
                data[v++]=j;
-               a=-((double*)(rec[0][ANG].data))[idx]*M_PI/180;
+               a=-((float*)(rec[0][ANG].data))[idx]*M_PI/180;
                angle_data[w++]=cos(a);
                angle_data[w++]=sin(a);
                for(n=0;n<g;n++){
@@ -111,7 +111,7 @@ int ReIndex(char **In,char *Out,char* FromTo,int *OtherDims,int BDW, int Orca) {
                if (in) {
                   data[v++]=i;
                   data[v++]=(FromTo[0]=='U')?j+rec[0][0].nj:j;
-                  a=-((double*)(rec[1][ANG].data))[idx]*M_PI/180;
+                  a=-((float*)(rec[1][ANG].data))[idx]*M_PI/180;
                   angle_data[w++]=cos(a);
                   angle_data[w++]=sin(a);
                   for(n=0;n<g;n++){
